Add support for including system-out and system-err
This change adds support for providing output and error lines to be
included in the XML output, similar to how attachments are
implemented. Since only one <system-out> tag may be present in the
XML, outputs and attachments are merged according to the JUnit
Attachments Plugin syntax - outputs first, then attachments, one per
line.

For #48