Testsuite time should include setup and teardown
The `time` attribute on `testsuite` should equal the amount of time
between the runner issuing the `suite` and `end suite` events.
It was previously calculated as the sum of each individual testcase's
`time`, but that doesn't account for setup and teardown.

Similarly, total time for all testsuites should be the amount of
time between the `start` and `end` events, which the `Base` reporter
stores in `stats.duration` for us.

Fixes #114